src/maintenance_server/files/home/requirements.txt
boto3
botocore
requests
pyyaml
troposphere
hvac
gitpython
paramiko
